# product-api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    Product:
        description: Product Product Product Product Product defines the structure for an API product
        x-go-package: example.com/hello/sdk/models
info:
    description: Documentation for Product API
    title: classification of Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Returns a list of products
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
        post:
            description: Create a product
            operationId: createProduct
            responses:
                "200":
                    $ref: '#/responses/createProductResponse'
            tags:
                - products
    /products/{id}:
        delete:
            description: Returns a list of products
            operationId: deleteProduct
            parameters:
                - description: ID of the product
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
            tags:
                - products
        get:
            description: Returns a single product
            operationId: getSingleProduct
            parameters:
                - description: ID of the product
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/createProductResponse'
            tags:
                - products
        put:
            description: Returns a list of products
            operationId: updateProduct
            parameters:
                - description: ID of the product
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/noContent'
            tags:
                - products
produces:
    - application/json
responses:
    createProductResponse:
        description: ""
    noContent:
        description: ""
    productsResponse:
        description: A list of products returns in the response
        schema:
            items: {}
            type: array
schemes:
    - http
swagger: "2.0"